Separate the web part from the rest without making major changes to a single file.
